Rename to EVV, with entry point evv, and package evv4esm
Because Eve is already a rather popular python package (a REST API for
humans) and has the namespace taken on pypi, we needed to rename EVE.
There is also discussion of developing a tool for  E3SM to access REST
APIs so there is a potential conflict there as well.

The package name will now be `evv4esm`, the common name (for docs and
discussion) will be `EVV` (still pronounced Eve),  and the entry point
will be `eve` (on the command line).